Update README and linebotservice file
This commit updates the README file and linebotservice file. The changes made to the README file include updating the installation instructions, adding a new section on troubleshooting, and fixing some typos. The linebotservice file was also modified to fix a bug related to user authentication. These changes will improve the overall functionality and user experience of the application.